+++ packages/cli/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
-## Recent Changes
+## `7.23.1`
 
 - Enhancement: Adding `--binary` and `--encoding` options to `zosfiles edit` to zowe V2